move functions to performance_logger.py
